chore: update akmods pubkey
As of May 19, 2023, in PR nvidia#104 a new akmods signing key was added
to the ublue github organization. The private key is held in org-level
secrets to enable use in multiple repos (eg, akmods and nvidia). This
add the pubkey here as it can be used immediately for building akmods.