feat: remove F42 builds for longterm-6.12 (#419)
The longterm-6.12 kernel was added for Cayo and now is planned for use in uCore. As it was always meant for server use it does not need the common kmod builds, except that uCore's build process does expect them for now.

It only needs to build for F43 now that Fedora CoreOS (and thus uCore) is on F43.