fix: ensure ZFS_MINOR_VERSION pin is respected
The ARG ZFS_MINOR_VERSION was too early in Containerfile, before the
last FROM statement before building ZFS kmod, thus it was ignored by
the RUN which built the kmod.

This moves it into the correct location and adds an echo to debug
what the provided version actually was.

Since zpools created on 2.4.0 can still be imported in 2.3.5
(verified in local testing), this will not only ensure we don't have
this occur again, but revert to 2.3.x for now.